Add links before heading paragraph mode
When the link mode is "paragraph", then intention is to get links at the
and of each reasonable section of the document. When sections don't end
in paragraphs, this did not behave in a useful manner. Solution: Always
add all accumulated links before printing the next heading and at the
end of the file.